app-controller/requirements.txt
# Core Framework
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.0
pyyaml>=6.0.1

# HTTP Client
httpx>=0.27.0
requests>=2.31.0

# GPU Monitoring
pynvml>=11.5.0

# Cache & Queue
redis>=5.0.1

# Monitoring & Metrics
prometheus-client>=0.20.0

# Image Processing
Pillow>=10.0.0

# System Monitoring
psutil>=5.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Code Quality (optional)
black>=24.0.0
flake8>=7.0.0
